rename package from bagelbids to bagel
