+++ README.md
@@ -1,8 +1,8 @@
-# Bagelbids
+# `bagel-cli`
 
-Bagelbids is a simple python command line tool to automatically read and annotate a 
+The Bagel CLI is a simple Python command line tool to automatically read and annotate a 
 [BIDS dataset](https://bids-specification.readthedocs.io/en/stable/) 
-so that it can be integrated into the neurobagel graph.
+so that it can be integrated into the Neurobagel graph.
 
 
 ## Dev environment
