+++ bagel/mappings.py
