+++ bagel/tests/data/example_synthetic.jsonld
