+++ bagel/tests/data/example_synthetic.tsv
