+++ bagel/utility.py
