+++ setup.cfg
@@ -45,7 +45,7 @@ all =
 
 [options.entry_points]
 console_scripts =
-    bagel=bagelbids.cli:bagel
+    bagel=bagel.cli:bagel
 
 [flake8]
 exclude =
